refactor: Unified the name to types.
